Integrate selenium-webdriver into cucumber for feature tests
I am not entirely sold on using cucumber-js instead of cucumber-rb; but
I think that it's more likely to be approachable for folks who are
coming from a react/node background and since that's where I will want
the most help I figure I may as well learn cucumber-js so that there's a
middle ground.

✅ Wire in Axios to hit the API
✅ Pass along data shaped like a Compost Pile
✨Connect When/Then for API steps to real API

~there is now an API endpoint for creating and showing CompostPile
objects.

While I only intend there to be one compost pile; having the ability to
create _moar_ makes it possible for as many tests to run in parallel as
the CI CPU/Memory can handle.

💡 Clarify purpose of the CompostPile model

- Added a TODO to clean up some weird code that's likely to be a sharp
  edge.

💡Do we need to specify email in the feature?

It's entirely possible that we will have the email address being generated randomly when we
build the Guest Sponsor for the test context.

If we specify the email in the feature, this may result in conflicts if we
were to run the tests in parallel.

If we rely on the email being informed from the test context, it may reduce
clarity of the feature file.

I think for now I would prefer to keep it out of the
feature file, and if we realize we need permutation
tests for the email address, we can cross that
bridge when we come to it.

So this is probably a no-op commit that will be useless later.

💡 Clean up the api and browser drivers and Cucumber steps

- Sprouted and Documented APIClient class
- Explored using jsdoc syntax
- Applied the default JavaScript linting / autoformatter

✨💄✅ Begin wiring the depositing compost tests and UI together

We took some time to learn a bit more about selenium, as well as to
start putting HTML into the page for the sponsorship form.

There's a few loose ends, in particular:

1. The "payment information" value, probably needs to be interpreted and
   spread across multiple input fields (like, their CC number, CVS, etc)
2. The cluster of promises to input fields feels like a place to inject
   a page or form object. See
    https://www.martinfowler.com/bliki/PageObject.html
3. We probably could make this the standard stripe checkout
   form: https://stripe.com/docs/payments/checkout/one-time