✅🔧 Remove cyclical dependency from testing infrastructure (#13)
Previously, the package.json file called `bin/test`, but `bin/test`
wants to call the individual sub-modules testing libraries. This could
result in recursion if I don't remember that the package.json was
referencing `bin/test`.

This also makes it more likely that someone who understands javascript
can just get started without worrying about the stuff in `bin`; because
they can start with reading the package.json to get an idea of what the
hell is going on in the project.