📝🔧  Generate API and System documentation from JavaScript comments (#14)
Documentation in public projects is particularly valuable because it allows
for contributions from a broader swathe of the development community.

Projects that have one person as the "source of truth" tend to struggle to survive past that persons zone of influence. Documentation helps expand the zone of influence and decrease the cost-of-contribution for folks.

That said, documentation can calso get out of date; which is not particularly great either. By relying on JSDoc, which reads it's documentation from the source code; we're more likely to keep the documentation and the implementation in sync.


* 🔧 Automate generation of jsdoc files

There are some weird things going on:

1. Links assume they are linking to markdown files, which breaks them
2. need to think more about how to organize /format the output so it can
   be web-accessible via github pages or the like.

* Make the documentation scripts names consistent